Move quick actions card related classes to cards package
